Events now support category, start and end dates.

+++ src/main/java/com/johannesbrodwall/infrastructure/db/Repository.java
@@ -6,7 +6,7 @@ public interface Repository<T> {
 
     List<T> findAll();
 
-    long insert(T entity);
+    int insert(T entity);
 
     T fetch(long id);
